mise en place du calendrie et evenement, update aprobation article, requette ajax

+++ espace_admin/delete.php
@@ -5,9 +5,16 @@ require('permission_admin.php');
 
 if(isset($_GET['id']) and !empty($_GET['id'])) {
     $getid = $_GET['id'];
+    
+    // Empêcher l'utilisateur de supprimer son propre compte
+    if ($getid == $_SESSION['id']) {
+        echo "Vous ne pouvez pas supprimer votre propre compte!";
+        exit;  // Arrête l'exécution du script
+    }
+
     $req = $conn->prepare('SELECT * FROM membres WHERE id = ?');
-    $req->execute(array($getid));//on vérifie si l'id existe dans la base de données
-    if($req->rowCount() > 0) {//si l'id existe
+    $req->execute(array($getid)); //on vérifie si l'id existe dans la base de données
+    if($req->rowCount() > 0) { //si l'id existe
         $delete = $conn->prepare('DELETE FROM membres WHERE id = ?');
         $delete->execute(array($getid));
         echo "Le membre a bien été supprimé";
@@ -15,7 +22,7 @@ if(isset($_GET['id']) and !empty($_GET['id'])) {
     } else {
         echo "Ce membre n'existe pas";
     }
-}else {
+} else {
     echo "L'id n'est pas recuperé";
 }
 ?>
